refactor(custody): isolate wallet provisioning
