+++ packages/sdp-custody/package.json
@@ -23,6 +23,11 @@
       "types": "./src/dfns/index.ts",
       "import": "./src/dfns/index.ts",
       "default": "./src/dfns/index.ts"
+    },
+    "./provisioning": {
+      "types": "./src/provisioning/index.ts",
+      "import": "./src/provisioning/index.ts",
+      "default": "./src/provisioning/index.ts"
     }
   },
   "scripts": {
@@ -42,6 +47,7 @@
     "@solana/keychain-utila": "1.3.0",
     "@solana/kit": "catalog:",
     "@solana/signers": "catalog:",
-    "@solana/transactions": "catalog:"
+    "@solana/transactions": "catalog:",
+    "jose": "^6.2.3"
   }
 }
